This is the trash file which is the most important in admin section
When the admin delete any think from the website it can not directly delete.
Becouse the institutes want to maintain and store thier data in database.So for this i have created this trash file so that all the out dated news and solved issues and feedback which we want to not show directly in the admin section can move here.